Automatically generate the SQL targetlist for classes.
Split DbStruct macro in two specific macros:

- CompareStruct that only implements the Compare trait*
- Dbstruct that also implements the Compare trait, and the new Sql trait

For now, the Sql trait only implement a tlist() function that returns the a Vec
of String describing the fields that should be returned for a given postgres
major version.  The App therefore retrieve the server_version_num GUC when
connecting, and is passed to every snapshot function.

Also the identifier for the database row stored in a DbStruct is now one of the
existing fields, defined in the DbStruct call, rather than the previous fixed
"ident" field.

The version requirement is optional in the DbStruct call.  If provided, the
underlying struct type will automatically be declared as Option<T> rather than
T, and the tlist() function will generate a proper value for that field for
unsupported version, of the form:

NULL::"field_type" AS "field_name"

Note that if the given version is PG_NO_CATALOG, then the struct field won't be
turned into an Option<T> and the tlist() function will ignore it.

The DbStruct now has the following syntax:

structname : ident_field {
    custom: type { PG_NO_CATALOG },
    mandatory: type,
    optional: type { minimum_pg_version },
    [...]
}

Note that there's no handling of maximum supported postgres version, but it
would be trivial to add if needed.