Implement a rust Char type as an alias for i8.
Postgres crate defines postgres char as an i8, so we have to add a lot of
useless code to output a char in the final diff as rust won't allow implement
FromSql for rust char.  Of course, it also mean that we can't use i8 for
anything else than postgres char.

The Char alias is only used to have a less surprising behavior, but of course
using i8 in the struct would behave the same.

This is implemented in a new pgtype file, as following commit(s) will add other
aliases to simplify the rust/postgres type translation.