Automatically instanciate a DbStruct from a postgres::Row
The Sql trait now has a now from_row() function, that instanciate a struct
from a given row, which is automatically implemented in the DbStruct macro.

This should work for the majority of the field that we care about.  The rest
will have to be manually implemented.

A first example of manual implementation is the list of attribute for a given
relation.  In order to do so, the Relation struct is split in two structs:

- PgClass: which implements the Compare and Sql traits and stores the pg_class
  atrributes we care about

- Relation: which only implements the Compare trait and stores:
  - a Vec<Attribute>
  - a PgClass

Note that splitting the struct would lead to multiple indirection error message
like:

Mismatch found for Relation XXX in PgClass:
Mismatch found for PgClass XXX in attributes:
[...]

To avoid that, the prefix "Pg" for struct names is reserved for this case, and
the two message will be merged in something like:

Mismatch found for Relation XXX in attributes:
[...]

As a consequence the PG_NO_CATALOG const is removed, as such fields should now
be in separate structs.