Add new rust/postgres type translations.
Doing so avoids to double check the postgres crate documentation for what
should be the proper rust definition of standard postgres datatypes.

It's also a good occasion to automatically generate the Compare trait for all
those aliases.  This is done with a new PgAlias macro.  A generic
implementation can't be done as it would conflict with HashMap traits, and
HashMap requires a dedicated implementation.