Support optional expression for DbStruct field declaration.
Some catalog fields have a misleading output (like pg_attribute.atttypmod),
and some other are totally unreadable (like pg_index.indexexprs), so simply
displaying those raw values wouldn't be really helpful to locate the
discrepancy in extension scripts.

A human readable output can be obtained using original postgres functions, but
it would require to split the catalog table struct in a CompareStruct with
those specific fields and a DbStruct for the rest of the fields, and executing
additional specific queries.

To avoid that, the DbStruct macro can now take an optional expression as a row
source instead of the original field name.  With this new syntax, retrieving a
human readable type name can be done with this syntax:

DbStruct!{
    Attribute:attname {
        attname: Name,
        atttype: Name = ("pg_catalog.format_type(a.atttypid, a.atttypmod)"),
    }
}

Note that rust syntax for expression isn't compatible with postgres one, so it
has to be defined as a plain literal to allow any SQL expression.