Handle extended statistics, and attach them to the owning table.
Also fix tlist() implementation to generate missing values as

NULL::type

This way, regular aliases work as expected and we can rely on e.g. integer
being resolved to int4.

There's on exception though.  For char, we need to explicitly double quote it
as it would otherwise be resolved to bpchar, which is not what we want.