Emit unified diff for multi-line text fields.
Those are emitted as a new SchemaDiff::UnifiedDiff enum value, which holds a
diffy::Patch, as provided by diffy.  Handling the Patch rather than its textual
representation is done in case other format are eventually added.

Note that diffy::Patch requires a lifetime, which needs to be added to the
Compare trait.   This is a good opportunity to change the SchemaDiff enums that
are compatible with a String slice rather than a plain String.

None of the currerntly handled fields would use this new format for now, but
this will be needed for the upcoming handling of routines.