Add a new proc_prototype! macro.
Retrieving a routine prototype requires a bit of SQL code, and it's already
duplicated one.  Later commits will need to retrieve a routine prototype for
other objects (at least for event triggers), so let's make it easier to
retrieve.

Note that doing so also fixes a bug in pg_proc.prosupport prototype, as it was
getting the list of arguments from the routine itself rather than the
prosupport routine.