Refactor BTreeMap and ClassOptions implementation for trait Compare.
The code was almost entirely duplicated, which isn't great.  A following commit
will implement a rust implementation for postgres arrays which shouldn't be
ordered also based on a BTreeMap, so duplicating the code a third time isn't an
option.

Instead create a new compare_map() function that implements the general
comparision logic, and takes a few parameters for the needs of the specific
implementation, including an optional function for the diffs accumulation.