Add a new List type, to store an unordered array of values.
The postgres array type is mapped to a Vec<T>, for which the element position
matters.  Some arrays in postgres catalogs hold a list of values for which the
order isn't important (for instant the list of command tags for an event
trigger), which can now be stored with this new List type.