Fix compatibility with PostgreSQL 13.
pg_range.rngmultitypid, pg_type.typsubscript and
pg_get_statisticsobjdef_columns() appeared in pg 14.

Also fix retrieval of pg_proc source.  All arms of the CASE expression will be
evaluated so it can't be used as a way to differentiate expression for the
various versions.

Thanks to Andrei Zubkov for the report!