Add suport for a configuration file.
All options can be specified in a configuration file.  Any option also
specified in the command line simply overrides the value specified in the
configuration file.

For now, two formats are supported: toml and json.  The format is automatically
detected using the file extension, and an error is reported if the extension is
not supported.