Provide a helpful error if DbStruct from_row() fails.
Until now, any mistake in the DbStruct metadata declaration could result in a
runtime panic, which is not friendly.  Instead intercept the problem and
provide as much information as possible to help diagnosis.

The error message is of the form:

Could not import PgClass row
relname: public.tbl0
column: reloptions
Error: error deserializing column 17: a Postgres value was `NULL`