Improve panic message for fallback value() function of trait Cmpare.
This function (and the nearby typname() one) is only provided to allow
compilation, as some struct implements the Compare trait but shouldn't never
have to provide a value.

As seen in af6f6c5ad2, it can be easy to overlook some codepath and realize
that either that function is called when it shouldn't be, or that a real
implementation should have been provided.  In any case, finding the root cause
wasn't especially easy with the current message, so improve it a bit and
provide at least the associated type in the message.