Detect GUC changes leaking out of the extension scrpts.
There's no  guarantee that the connection used to create the extension won't be
kept to do some other work, so it's important that no configuration change
leaks out of the script.

Note that this is a best effort only.  This tool can't detect a script doing a
RESET [ guc_name | ALL ], which could also revert any change done in the
connection or transaction before creating the extension.