Fix operator family identifiers.
The previous code was broken, as an operator family name is only unique for a
given access method.

Thanks to Zacharias Knudsen for the report.  As seen in the bug report, it can
lead to false positive problem as the comparision was depending on the physical
tuple positions in the table.

Tests added to reproduce the problem in the pg_broken_extupgrade extension in
"head" version.  I wasn't able to reproduce it with only 2 operator families
though.  There is also another test to cover an actual diff with operator
families.