Fix compatibility with pg 9.5 and lower
The automatic tlist implementation doesn't cope well with fields defined as
Option, as it relies on stringify to generate the target SQL type.

For now just intercept any field defined as Option<Text>, as the only problems
happens in pg_aggregate.rs on pg 9.5 and below, but I should eventually find a
better solution.